+++ openhands/tools/glob/definition.py
@@ -98,7 +98,7 @@ class GlobTool(ToolDefinition[GlobAction, GlobObservation]):
                          conv_state.workspace
         """
         # Import here to avoid circular imports
-        from openhands.tools.file_glob.impl import GlobExecutor
+        from openhands.tools.glob.impl import GlobExecutor
 
         working_dir = conv_state.workspace.working_dir
         if not os.path.isdir(working_dir):
