+++ openhands/tools/preset/planning.py
@@ -12,7 +12,7 @@ logger = get_logger(__name__)
 
 def register_planning_tools() -> None:
     """Register the planning agent tools."""
-    from openhands.tools.file_glob import GlobTool
+    from openhands.tools.glob import GlobTool
     from openhands.tools.grep import GrepTool
     from openhands.tools.planning_file_editor import PlanningFileEditorTool
